add email validator package, update request validation for nested object and move to untils folder
