+++ package.json
@@ -40,7 +40,8 @@
         "stream-browserify": "^3.0.0",
         "stream-http": "^3.2.0",
         "url": "^0.11.3",
-        "util": "^0.12.5"
+        "util": "^0.12.5",
+        "validator": "^13.12.0"
     },
     "dependencies": {
         "@fortawesome/fontawesome-free": "^6.5.2",
